requirements.txt
Flask~=0.10.1
docker-py~=1.3.1
